Make sure JSON files aren't recognized as WebP
Previously WebP parser checked for a plain string in middle of a file,
enabling possible vulnerabilities like uploading of JSON files when
only images are meant to be uploaded.

Considering WebP files are stored within RIFF containers, it makes sense
to check whether a WebP file is stored in such container.